requirements.txt
faiss-cpu
torch
tensorflow
tensorflow-hub
transformers
Flask
